+++ Templates/settings.blade.php
@@ -13,15 +13,8 @@
         <div class="maincontentinner timetable-settings">
             <h5 class="subtitle">TimeTable Settings</h5>
             <p class="tw-pb-m">These settings will change the way the TimeTable plugin works.</p>
-            <h4 class="widgettitle title-light"><span class="fa fa-cog"></span>
-                Cache expiration durations
-            </h4>
+
             <form method="post" id="" action="<?= BASE_URL ?>/TimeTable/settings">
-                <div class="row">
-                    <div class="col-md-2">
-                        <label>Tickets (minutes)</label>
-                    </div>
-                </div>
                 <div class="row">
                     <div class="col-md-2">
                         <label>Require timelog comment</label>
@@ -31,6 +24,7 @@
                             {{ (int) $requireTimeRegistrationComment === 1 ? 'checked' : '' }} />
                     </div>
                 </div>
+                <br>
                 <input type="submit" value="Save" id="saveBtn" />
             </form>