+++ Templates/timetable.blade.php
@@ -105,10 +105,10 @@
                                             $id = $timesheetDate[0]['id'] ?? null;
                                             $headline = $timesheetDate[0]['headline'] ?? null;
                                             $hours = $timesheetDate[0]['hours'] ?? null;
-                                            $hoursLeft = $timesheetDate[0]['hourRemaining'] ?? null;
+                                            $hoursLeft = $timesheet['hourRemaining'] ?? null;
                                             $description = $timesheetDate[0]['description'] ?? null;
                                             $requireTimeRegistrationComment = $requireTimeRegistrationComment ?? 0;
-                                            $isMissingDescription = isset($hours) & (trim($description) === '') && $requireTimeRegistrationComment !== 0;
+                                            $isMissingDescription = isset($hours) & (trim($description) === '') && $requireTimeRegistrationComment !== '0';
                                             
                                             // accumulate hours
                                             if ($hours) {